Django_restaurant/pyproject.toml
[tool.poetry]
name = "django-restaurant"
version = "0.1.0"
description = ""
authors = ["woojin-an <dkswoojin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-restframework = "^0.0.1"
pymysql = "^1.1.1"
pillow = "^10.4.0"
django-cleanup = "^8.1.0"
django-extensions = "^3.2.3"
mysqlclient = "^2.2.4"
djangorestframework = "^3.15.2"
serializers = "^0.2.4"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
